requirements.txt
gym~=0.26.2
gymnasium~=0.29.1
pandas~=1.5.3
pyYAML~=5.3.1
tqdm~=4.65.0
wandb~=0.16.0
joblib~=1.1.0
numpy~=1.25.1
pyglet~=1.5.15
setuptools~=58.0.4
scipy~=1.11.4
matplotlib~=3.4.3
seaborn~=0.13.0
optuna